Refactor local and dome charging engines to use applied billing rates for local processing
